requirements.txt
streamlit==1.27.2
opencv-python-headless==4.7.0.72
mediapipe==0.9.1.0
numpy==1.23.5
pandas==1.5.3
pillow==9.5.0
matplotlib==3.7.1
protobuf==3.20.1
attrs==23.1.0